Move useEffect inside useActionHooks.
